:heavy_plus_sign: Add Users class
